Some fixes for OS X

+++ building/vim.sh
@@ -11,15 +11,24 @@ vim_repo='https://api.github.com/repos/vim/vim/tags'
 
 echo getting latest vim version:
 # FROM: "name": "v8.1.0338",
-# TO:   8.1.0338
-version=$(curl --silent $vim_repo | grep name | head -n1 | sed -r 's/.*v([^"]+)".*/\1/')
+# TO:   v8.1.0338
+version=$(curl --silent $vim_repo | grep name | head -n1 | sed 's/.*: "\(.*\)".*/\1/')
+
 echo found: $version
+
+echo looking for python3
+if ! which python3; then
+    echo You must have python3 installed otherwise the vim
+    echo configure step would silently avoid enabling support
+    echo for it in the vim build.
+    exit 1
+fi
 
 cd /tmp
 
 [[ -d vim ]] && rm -rf vim
 
-git clone --depth=1 --branch=v$version https://github.com/vim/vim
+git clone --depth=1 --branch=$version https://github.com/vim/vim
 
 cd vim
 
@@ -41,5 +50,10 @@ rm -f ~/bin/vimdiff
 
 mkdir -p ~/bin
 
-ln -s $prefix/bin/vim ~/bin/vim
-ln -s $prefix/bin/vimdiff ~/bin/vimdiff
+pushd "$tools"
+rm -f vim-current
+ln -s vim-$version vim-current
+popd
+
+ln -s $tools/vim-current/bin/vim     ~/bin/vim
+ln -s $tools/vim-current/bin/vimdiff ~/bin/vimdiff
